+++ adapter_test/support/identity_case.ex
@@ -36,6 +36,7 @@ if Application.compile_env!(:ithibati, :probe_adapter) in [
 
       Ecto.Migrator.up(Repo, 10, Ithibati.AdapterApplicationMigration, log: false)
       Ecto.Migrator.up(Repo, 11, Ithibati.AdapterLibraryMigration, log: false)
+      Ecto.Migrator.up(Repo, 14, Ithibati.AdapterChallengeMigration, log: false)
       clear()
       on_exit(&clear/0)
       :ok
@@ -115,6 +116,7 @@ if Application.compile_env!(:ithibati, :probe_adapter) in [
     end
 
     defp clear do
+      Repo.delete_all(Ithibati.Challenge)
       Repo.delete_all(Ithibati.Bootstrap)
       Repo.delete_all(Ithibati.AdapterUser)
       Repo.delete_all(Ithibati.AdapterInvitation)