+++ test/support/migrations/20260918000000_add_challenges.exs
@@ -1,0 +1,5 @@
+defmodule Ithibati.TestRepo.Migrations.AddChallenges do
+  use Ecto.Migration
+  def up, do: Ithibati.Migration.up(from: 1, version: 2)
+  def down, do: Ithibati.Migration.down(from: 1, version: 2)
+end